Create option for simple cfc completion names
See #57 - local buffer completions are blocked if word separators are
included in completion names. This can be worked around by using a
plugin such as `All Autocomplete`, but it seems prudent to also include
an option in the package for completion names that do not block ST's
local buffer completions.